Script to generate report from output of gst-launch-start-stop.sh
There are a lot of global variables and strong coupling (e.g. between
the "ts" (timestamp) and "search" functions, because "ts" assumes that
"m" has been set to a MatchObject with a timestamp in the first group).
This greatly simplified the parser so that hopefully all you need is to
understand the state_machine table.

It would probably have been easier & more natural to write this in perl.

The "-*- eval..." line is an Emacs magic comment to disable
pretty-symbols-mode[1]. In my Emacs configuration I enable
pretty-symbols-mode automatically so that the word "lambda" is displayed
as "λ". But this ruins the formatting of the state_machine table, so I
want it disabled just for this file. It won't hurt non-Emacs users
because it's inside a python comment.

[1] https://github.com/drothlis/pretty-symbols